# docs/testing/api/openapi.docker.yml
openapi: 3.1.0
info:
  title: Reservas Médicas - Auth API
  version: 1.0.0
  description: OpenAPI spec mínimo para pruebas de seguridad (endpoints expuestos en server/index.js)
servers:
  - url: http://host.docker.internal:4000
paths:
  /auth/login:
    post:
      summary: Inicia sesión (demo)
      description: >
        Endpoint de login demo: recibe { email, password } y devuelve { token, user }.
        En el servidor se valida contra un usuario demo y se retorna un JWT con 1h de vida.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: demo@med.app
                password:
                  type: string
                  example: '123456'
      responses:
        '200':
          description: Login exitoso - devuelve token y datos del usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'Usuario Demo'
                      email:
                        type: string
                        example: demo@med.app
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Credenciales inválidas'
  /profile:
    get:
      summary: Perfil del usuario (protegido)
      description: "Devuelve la información del usuario autenticado. Requiere header 'Authorization: Bearer <token>'."
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Datos del perfil
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  email:
                    type: string
                    example: demo@med.app
                  name:
                    type: string
                    example: 'Usuario Demo'
        '401':
          description: Token requerido / inválido / expirado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Token requerido'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT